requirements.txt
# Core dependencies
requests==2.31.0
urllib3==2.1.0
tqdm>=4.66.0
jira-python

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# AI/ML
google-generativeai>=0.8.0
chromadb>=0.5.0
numpy>=1.22.0  # Compatible with latest ChromaDB

# Image Processing
Pillow==10.2.0
pytesseract
